+++ data/auras.lua
@@ -3,9 +3,11 @@ local _, Engine = ...
 -- This database will contain lists of categorized Auras
 -- for CC, dispellers and so on. 
 
-
-
 local auras = {
+	-- whitelisted auras mainly for the target frame
+	whitelist = {
+
+	},
 	cc = {},
 	harm = {},
 	help = {},